backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
PyJWT>=2.8.0

# AI & ML
google-generativeai>=0.8.0
python-dotenv==1.0.0
python-multipart>=0.0.6

# External APIs
google-api-python-client>=2.100.0
requests>=2.31.0
youtube-transcript-api>=0.6.0
PyPDF2>=3.0.0
PyMuPDF>=1.26.0

# CORS & Middleware
fastapi-cors>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Utilities
python-dateutil>=2.8.2
email-validator>=2.1.0
